Chore: Rename SnsTransactionModal in neurons
